Move urban_idea_special_case to main.rs
We need to break up obsidianize, because parse needs to know the file
name to do its job.

Split `obsidianize` into `subdivide` and `parse`

This is because `parse` needs to know the output file name to create
Obsidian links, and the output file name depends on both the external
file name, known by `main`, and the internal heading that `subdivide`
returns.