Improvements for some file names
Obsidian file names serve as note titles, displayed in the note body and
also the table of contents in the left sidebar

We prefer the Header 1 title in the file body to the file-system name,
unless the file-system name is longer. (Some of the Laironomicon files
have a Header 1 of "Name", so we use the more complete name
contained in the copyright line (and is better than the original file
name).

All this results in just one file with a number >= 100 (100 Brain
Devourer’s Domain in the Laironomicon); rather than using three digits
for every number, we choose to remove the number entirely and use "Brain
Devourer's Domain", which sorts to the end of the list as it should.