Convert compare_to_baseline from a script to a test
We now need to compare output to baseline when created from the zipfile
source as well as when created from the directory source, and I know
Rust better than bash. So a test in tests/compare_to_baseline.rs
supercedes compare-to-baseline.sh.

The test is a bit long, and we can't run it in CI because it relies on
copyrighted material, so we use #[ignore].

Since .git/hooks/pre-push has a line

    cargo test -- --ignored

it will be tested before we push to github.

The compare_to_baseline test ensures that a refacoring truly doesn't
change behavior.  For intended changes, the script update_baseline.sh
copies the latest output to baseline after checking that the output from
zipfile input and the output from directory input are identical.