Add trailing vertical bars to each table Lines
When we do this, Obsidian displays the rendered table in editing mode.

When we don't do this, Obsidian displays the Markdown code for the table in
editing mode, and the rendered table only in reading mode.

(And it occurs to me that table plugins might well expect the trailing
vertical bars.)